Language.Haskell.GhcMod

Description

The ghc-mod library.

Cradle

data Cradle

The environment where this library is used.

Constructors

Cradle

Fields

cradleProject :: Project

cradleCurrentDir :: FilePath

The directory where this library is executed.

cradleRootDir :: FilePath

The project root directory.

cradleTempDir :: FilePath

Per-Project temporary directory

cradleCabalFile :: Maybe FilePath

The file name of the found cabal file.

cradleDistDir :: FilePath

The build info directory.

data Project

Constructors

CabalProject

SandboxProject

PlainProject

StackProject StackEnv

findCradle :: (IOish m, GmOut m) => m Cradle

Finding Cradle. Find a cabal file by tracing ancestor directories. Find a sandbox according to a cabal sandbox config in a cabal directory.

Options

data Options

Constructors

Options

Fields

optOutput :: OutputOpts

optPrograms :: Programs

optGhcUserOptions :: [GHCOption]

GHC command line options set on the ghc-mod command line

optOperators :: Bool

If True, browse also returns operators.

optDetailed :: Bool

If True, browse also returns types.

optQualified :: Bool

If True, browse will return fully qualified name

optHlintOpts :: [String]

optFileMappings :: [(FilePath, Maybe FilePath)]

newtype LineSeparator

The type for line separator. Historically, a Null string is used.

Constructors

LineSeparator String

data OutputStyle

Output style.

Constructors

LispStyle
S expression style.

PlainStyle
Plain textstyle.

data FileMapping

Constructors

FileMapping

Fields

fmPath :: FilePath

fmTemp :: Bool

defaultOptions :: Options

A default Options.

Logging

data GmLogLevel

increaseLogLevel :: GmLogLevel -> GmLogLevel

decreaseLogLevel :: GmLogLevel -> GmLogLevel

gmSetLogLevel :: GmLog m => GmLogLevel -> m ()

gmLog :: (MonadIO m, GmLog m, GmOut m) => GmLogLevel -> String -> Doc -> m ()

>>> Just GmDebug <= Nothing
False
>>> Just GmException <= Just GmDebug
True
>>> Just GmDebug <= Just GmException
False

Types

data ModuleString

Module name.

newtype Expression

Haskell expression.

Constructors

Expression

Fields

getExpression :: String

data GhcPkgDb

GHC package database flags.

type Symbol = String

Type of function and operation names.

data SymbolDb

Database from Symbol to [ModuleString].

data GhcModError

Constructors

GMENoMsg
Unknown error

GMEString String
Some Error with a message. These are produced mostly by fail calls on GhcModT.

GMECabalConfigure GhcModError
Configuring a cabal project failed.

GMEStackConfigure GhcModError
Configuring a stack project failed.

GMEStackBootstrap GhcModError
Bootstrapping stack environment failed (process exited with failure)

GMECabalCompAssignment [(Either FilePath ModuleName, Set ChComponentName)]
Could not find a consistent component assignment for modules

GMEProcess String String [String] (Either Int GhcModError)
Launching an operating system process failed. Fields in order: function, command, arguments, (stdout, stderr, exitcode)

GMENoCabalFile
No cabal file found.

GMETooManyCabalFiles [FilePath]
Too many cabal files found.

GMEWrongWorkingDirectory FilePath FilePath

Monad Types

type GhcModT m = GmT (GmOutT m)

type IOish m = (Functor m, MonadIO m, MonadBaseControl IO m, ExceptionMonad m)

A constraint alias (-XConstraintKinds) to make functions dealing with GhcModT somewhat cleaner.

Basicially an IOish m => m is a Monad supporting arbitrary IO and exception handling. Usually this will simply be IO but we parametrise it in the exported API so users have the option to use a custom inner monad.

Monad utilities

runGhcModT :: (IOish m, GmOut m) => Options -> GhcModT m a -> m (Either GhcModError a, GhcModLog)

Run a GhcModT m computation.

withOptions :: GmEnv m => (Options -> Options) -> m a -> m a

dropSession :: IOish m => GhcModT m ()

Drop the currently active GHC session, the next that requires a GHC session will initialize a new one.

GhcMod utilities

boot :: IOish m => GhcModT m String

Printing necessary information for front-end booting.

browse

Arguments

:: IOish m

=> String
A module name. (e.g. "Data.List", "base:Prelude")

-> GhcModT m String

Getting functions, classes, etc from a module. If detailed is True, their types are also obtained. If operators is True, operators are also returned.

check

Arguments

:: IOish m

=> [FilePath]
The target files.

-> GhcModT m (Either String String)

Checking syntax of a target file using GHC. Warnings and errors are returned.

checkSyntax

Arguments

:: IOish m

=> [FilePath]
The target files.

-> GhcModT m String

Checking syntax of a target file using GHC. Warnings and errors are returned.

debugInfo :: IOish m => GhcModT m String

Obtaining debug information.

componentInfo :: IOish m => [String] -> GhcModT m String

expandTemplate

Arguments

:: IOish m

=> [FilePath]
The target files.

-> GhcModT m String

Expanding Haskell Template.

info

Arguments

:: IOish m

=> FilePath
A target file.

-> Expression
A Haskell expression.

-> GhcModT m String

Obtaining information of a target expression. (GHCi's info:)

lint

Arguments

:: IOish m

=> FilePath
A target file.

-> GhcModT m String

Checking syntax of a target file using hlint. Warnings and errors are returned.

pkgDoc :: IOish m => String -> GhcModT m String

Obtaining the package name and the doc path of a module.

rootInfo :: (IOish m, GmOut m) => m String

Obtaining root information.

types

Arguments

:: IOish m

=> FilePath
A target file.

-> Int
Line number.

-> Int
Column number.

-> GhcModT m String

Obtaining type of a target expression. (GHCi's type:)

splits

Arguments

:: IOish m

=> FilePath
A target file.

-> Int
Line number.

-> Int
Column number.

-> GhcModT m String

Splitting a variable in a equation.

sig

Arguments

:: IOish m

=> FilePath
A target file.

-> Int
Line number.

-> Int
Column number.

-> GhcModT m String

Create a initial body from a signature.

refine

Arguments

:: IOish m

=> FilePath
A target file.

-> Int
Line number.

-> Int
Column number.

-> Expression
A Haskell expression.

-> GhcModT m String

auto

Arguments

:: IOish m

=> FilePath
A target file.

-> Int
Line number.

-> Int
Column number.

-> GhcModT m String

modules :: (IOish m, Gm m) => m String

Listing installed modules.

languages :: IOish m => GhcModT m String

Listing language extensions.

flags :: IOish m => GhcModT m String

Listing GHC flags. (e.g -fno-warn-orphans)

findSymbol :: IOish m => Symbol -> GhcModT m String

Looking up SymbolDb with Symbol to [ModuleString] which will be concatenated. loadSymbolDb is called internally.

lookupSymbol :: IOish m => Symbol -> SymbolDb -> GhcModT m String

Looking up SymbolDb with Symbol to [ModuleString] which will be concatenated.

dumpSymbol :: IOish m => FilePath -> GhcModT m String

Dumping a set of (Symbol,[ModuleString]) to a file if the file does not exist or is invalid. The file name is printed.

SymbolDb

loadSymbolDb :: IOish m => FilePath -> GhcModT m SymbolDb

Loading a file and creates SymbolDb.

isOutdated :: IOish m => SymbolDb -> GhcModT m Bool

Output

gmPutStr :: (MonadIO m, GmOut m) => String -> m ()

gmErrStr :: (MonadIO m, GmOut m) => String -> m ()

gmPutStrLn :: (MonadIO m, GmOut m) => String -> m ()

gmErrStrLn :: (MonadIO m, GmOut m) => String -> m ()

FileMapping

loadMappedFile :: IOish m => FilePath -> FilePath -> GhcModT m ()

loadMappedFileSource :: IOish m => FilePath -> String -> GhcModT m ()

unloadMappedFile :: IOish m => FilePath -> GhcModT m ()