Language.Haskell.GhcMod.Error

Documentation

data GhcModError

Constructors

GMENoMsg
Unknown error

GMEString String
Some Error with a message. These are produced mostly by fail calls on GhcModT.

GMECabalConfigure GhcModError
Configuring a cabal project failed.

GMEStackConfigure GhcModError
Configuring a stack project failed.

GMEStackBootstrap GhcModError
Bootstrapping stack environment failed (process exited with failure)

GMECabalCompAssignment [(Either FilePath ModuleName, Set ChComponentName)]
Could not find a consistent component assignment for modules

GMEProcess String String [String] (Either Int GhcModError)
Launching an operating system process failed. Fields in order: function, command, arguments, (stdout, stderr, exitcode)

GMENoCabalFile
No cabal file found.

GMETooManyCabalFiles [FilePath]
Too many cabal files found.

GMEWrongWorkingDirectory FilePath FilePath

type GmError m = MonadError GhcModError m

gmeDoc :: GhcModError -> Doc

ghcExceptionDoc :: GhcException -> Doc

liftMaybe :: MonadError e m => e -> m (Maybe a) -> m a

overrideError :: MonadError e m => e -> m a -> m a

modifyError :: MonadError e m => (e -> e) -> m a -> m a

modifyError' :: MonadError e m => m a -> (e -> e) -> m a infixr 0

modifyGmError :: (MonadIO m, ExceptionMonad m) => (GhcModError -> GhcModError) -> m a -> m a

tryFix :: MonadError e m => m a -> (e -> m ()) -> m a

data GHandler m a

Constructors

forall e . Exception e => GHandler (e -> m a)

gcatches :: (MonadIO m, ExceptionMonad m) => m a -> [GHandler m a] -> m a

module Control.Monad.Error

module Control.Exception