Language.Haskell.GhcMod.GhcPkg

Documentation

ghcPkgDbOpt :: GhcPkgDb -> [String]

ghcPkgDbStackOpts

Arguments

:: [GhcPkgDb]
Package db stack

-> [String]

Get options needed to add a list of package dbs to ghc-pkg's db stack

ghcDbStackOpts

Arguments

:: [GhcPkgDb]
Package db stack

-> [String]

Get options needed to add a list of package dbs to ghc's db stack

ghcDbOpt :: GhcPkgDb -> [String]

getPackageDbStack :: IOish m => GhcModT m [GhcPkgDb]

getPackageCachePaths :: IOish m => FilePath -> GhcModT m [FilePath]

getGhcPkgProgram :: IOish m => GhcModT m FilePath