Language.Haskell.GhcMod.Internal

Description

Low level access to the ghc-mod library.

Types

type GHCOption = String

A single GHC command line option.

type Package = (PackageBaseName, PackageVersion, PackageId)

A package's name, verson and id.

type PackageBaseName = String

A package name.

type PackageVersion = String

A package version.

type PackageId = String

A package id.

type IncludeDir = FilePath

An include directory for modules.

newtype GmlT m a

Constructors

GmlT

Fields

unGmlT :: GhcModT m a

class MonadIOC m => MonadIO m where

Methods

liftIO :: IO a -> m a

class Monad m => GmEnv m where

Minimal complete definition

(gmeAsk | gmeReader), gmeLocal

Methods

gmeAsk :: m GhcModEnv

gmeReader :: (GhcModEnv -> a) -> m a

gmeLocal :: (GhcModEnv -> GhcModEnv) -> m a -> m a

Various Paths

ghcLibDir :: FilePath

Obtaining the directory for ghc system libraries.

ghcModExecutable :: IO FilePath

Returns the path to the currently running ghc-mod executable. With ghc<7.6 this is a guess but >=7.6 uses getExecutablePath.

Logging

withLogger :: (GmGhc m, GmEnv m, GmOut m, GmState m) => (DynFlags -> DynFlags) -> m a -> m (Either String (String, a))

Logged messages are returned as String. Right is success and Left is failure.

setNoWarningFlags :: DynFlags -> DynFlags

Set DynFlags equivalent to "-w:".

setAllWarningFlags :: DynFlags -> DynFlags

Set DynFlags equivalent to "-Wall".

Environment, state and logging

data GhcModEnv

Constructors

GhcModEnv

Fields

gmOptions :: Options

gmCradle :: Cradle

data GhcModState

data CompilerMode

Constructors

Simple

Intelligent

data GhcModLog

class Monad m => GmLog m where

Methods

gmlJournal :: GhcModLog -> m ()

gmlHistory :: m GhcModLog

gmlClear :: m ()

data GmLogLevel

Constructors

GmSilent

GmPanic

GmException

GmError

GmWarning

GmInfo

GmDebug

GmVomit

gmSetLogLevel :: GmLog m => GmLogLevel -> m ()

Monad utilities

runGhcModT' :: IOish m => GhcModEnv -> GhcModState -> GhcModT m a -> GmOutT m (Either GhcModError (a, GhcModState), GhcModLog)

Run a computation inside GhcModT providing the RWST environment and initial state. This is a low level function, use it only if you know what to do with GhcModEnv and GhcModState.

You should probably look at runGhcModT instead.

hoistGhcModT :: IOish m => (Either GhcModError a, GhcModLog) -> GhcModT m a

hoistGhcModT result. Embed a GhcModT computation's result into a GhcModT computation. Note that if the computation that returned result modified the state part of GhcModT this cannot be restored.

runGmlT :: IOish m => [Either FilePath ModuleName] -> GmlT m a -> GhcModT m a

runGmlT' :: IOish m => [Either FilePath ModuleName] -> (DynFlags -> Ghc DynFlags) -> GmlT m a -> GhcModT m a

gmlGetSession :: (MonadIO m, MonadBaseControl IO m) => GmlT m HscEnv

gmlSetSession :: (MonadIO m, MonadBaseControl IO m) => HscEnv -> GmlT m ()

loadTargets :: IOish m => [GHCOption] -> [FilePath] -> GmlT m ()

Set the files as targets and load them.

cabalResolvedComponents :: IOish m => GhcModT m (Map ChComponentName (GmComponent GMCResolved (Set ModulePath)))

Accessing GhcModEnv and GhcModState

options :: GmEnv m => m Options

cradle :: GmEnv m => m Cradle

getCompilerMode :: GmState m => m CompilerMode

setCompilerMode :: GmState m => CompilerMode -> m ()

targetGhcOptions :: forall m. IOish m => Cradle -> Set (Either FilePath ModuleName) -> GhcModT m [GHCOption]

withOptions :: GmEnv m => (Options -> Options) -> m a -> m a

GhcModError

gmeDoc :: GhcModError -> Doc

World

data World

getCurrentWorld :: IOish m => GhcModT m World

didWorldChange :: IOish m => World -> GhcModT m Bool

Cabal Helper

data ModulePath

Constructors

ModulePath

Fields

mpModule :: ModuleName

mpPath :: FilePath

data GmComponent t eps

Constructors

GmComponent

Fields

gmcHomeModuleGraph :: GmModuleGraph

gmcName :: ChComponentName

gmcGhcOpts :: [GHCOption]

gmcGhcPkgOpts :: [GHCOption]

gmcGhcSrcOpts :: [GHCOption]

gmcGhcLangOpts :: [GHCOption]

gmcRawEntrypoints :: ChEntrypoint

gmcEntrypoints :: eps

gmcSourceDirs :: [FilePath]

data GmComponentType

Constructors

GMCRaw

GMCResolved

data GmModuleGraph

Constructors

GmModuleGraph

Fields

gmgGraph :: Map ModulePath (Set ModulePath)

prepareCabalHelper :: (IOish m, GmEnv m, GmOut m, GmLog m) => m ()

Misc stuff

data GHandler m a

Constructors

forall e . Exception e => GHandler (e -> m a)

gcatches :: (MonadIO m, ExceptionMonad m) => m a -> [GHandler m a] -> m a

FileMapping

module Language.Haskell.GhcMod.FileMapping