Language.Haskell.GhcMod.Monad.Types

Monad Types

type GhcModT m = GmT (GmOutT m)

newtype GmOutT m a

Constructors

GmOutT

Fields

unGmOutT :: ReaderT GhcModOut m a

newtype GmT m a

Constructors

GmT

Fields

unGmT :: StateT GhcModState (ErrorT GhcModError (JournalT GhcModLog (ReaderT GhcModEnv m))) a

newtype GmlT m a

Constructors

GmlT

Fields

unGmlT :: GhcModT m a

newtype LightGhc a

Constructors

LightGhc

Fields

unLightGhc :: ReaderT (IORef HscEnv) IO a

type GmGhc m = (IOish m, GhcMonad m)

type IOish m = (Functor m, MonadIO m, MonadBaseControl IO m, ExceptionMonad m)

A constraint alias (-XConstraintKinds) to make functions dealing with GhcModT somewhat cleaner.

Basicially an IOish m => m is a Monad supporting arbitrary IO and exception handling. Usually this will simply be IO but we parametrise it in the exported API so users have the option to use a custom inner monad.

Environment, state and logging

data GhcModEnv

Constructors

GhcModEnv

Fields

gmOptions :: Options

gmCradle :: Cradle

data GhcModState

Constructors

GhcModState

Fields

gmGhcSession :: !(Maybe GmGhcSession)

gmComponents :: !(Map ChComponentName (GmComponent GMCResolved (Set ModulePath)))

gmCompilerMode :: !CompilerMode

gmCaches :: !GhcModCaches

gmMMappedFiles :: !FileMappingMap

data GhcModCaches

Constructors

GhcModCaches

Fields

gmcPackageDbStack :: CacheContents ChCacheData [GhcPkgDb]

gmcMergedPkgOptions :: CacheContents ChCacheData [GHCOption]

gmcComponents :: CacheContents ChCacheData [GmComponent GMCRaw ChEntrypoint]

gmcResolvedComponents :: CacheContents [GmComponent GMCRaw (Set ModulePath)] (Map ChComponentName (GmComponent GMCResolved (Set ModulePath)))

defaultGhcModState :: GhcModState

data GmGhcSession

Constructors

GmGhcSession

Fields

gmgsOptions :: ![GHCOption]

gmgsSession :: !(IORef HscEnv)

data GmComponent t eps

Constructors

GmComponent

Fields

gmcHomeModuleGraph :: GmModuleGraph

gmcName :: ChComponentName

gmcGhcOpts :: [GHCOption]

gmcGhcPkgOpts :: [GHCOption]

gmcGhcSrcOpts :: [GHCOption]

gmcGhcLangOpts :: [GHCOption]

gmcRawEntrypoints :: ChEntrypoint

gmcEntrypoints :: eps

gmcSourceDirs :: [FilePath]

data CompilerMode

Constructors

Simple

Intelligent

Accessing GhcModEnv, GhcModState and GhcModLog

data GmLogLevel

Constructors

GmSilent

GmPanic

GmException

GmError

GmWarning

GmInfo

GmDebug

GmVomit

data GhcModLog

Constructors

GhcModLog

Fields

gmLogLevel :: Maybe GmLogLevel

gmLogVomitDump :: Last Bool

gmLogMessages :: [(GmLogLevel, String, Doc)]

data GhcModError

Constructors

GMENoMsg
Unknown error

GMEString String
Some Error with a message. These are produced mostly by fail calls on GhcModT.

GMECabalConfigure GhcModError
Configuring a cabal project failed.

GMEStackConfigure GhcModError
Configuring a stack project failed.

GMEStackBootstrap GhcModError
Bootstrapping stack environment failed (process exited with failure)

GMECabalCompAssignment [(Either FilePath ModuleName, Set ChComponentName)]
Could not find a consistent component assignment for modules

GMEProcess String String [String] (Either Int GhcModError)
Launching an operating system process failed. Fields in order: function, command, arguments, (stdout, stderr, exitcode)

GMENoCabalFile
No cabal file found.

GMETooManyCabalFiles [FilePath]
Too many cabal files found.

GMEWrongWorkingDirectory FilePath FilePath

type Gm m = (GmEnv m, GmState m, GmLog m, GmOut m)

class Monad m => GmEnv m where

Minimal complete definition

(gmeAsk | gmeReader), gmeLocal

Methods

gmeAsk :: m GhcModEnv

gmeReader :: (GhcModEnv -> a) -> m a

gmeLocal :: (GhcModEnv -> GhcModEnv) -> m a -> m a

class Monad m => GmState m where

Minimal complete definition

gmsState | gmsGet, gmsPut

Methods

gmsGet :: m GhcModState

gmsPut :: GhcModState -> m ()

gmsState :: (GhcModState -> (a, GhcModState)) -> m a

class Monad m => GmLog m where

Methods

gmlJournal :: GhcModLog -> m ()

gmlHistory :: m GhcModLog

gmlClear :: m ()

class Monad m => GmOut m where

Methods

gmoAsk :: m GhcModOut

cradle :: GmEnv m => m Cradle

options :: GmEnv m => m Options

outputOpts :: GmOut m => m OutputOpts

withOptions :: GmEnv m => (Options -> Options) -> m a -> m a

getCompilerMode :: GmState m => m CompilerMode

setCompilerMode :: GmState m => CompilerMode -> m ()

getMMappedFiles :: GmState m => m FileMappingMap

setMMappedFiles :: GmState m => FileMappingMap -> m ()

addMMappedFile :: GmState m => FilePath -> FileMapping -> m ()

delMMappedFile :: GmState m => FilePath -> m ()

lookupMMappedFile :: GmState m => FilePath -> m (Maybe FileMapping)

getMMappedFilePaths :: GmState m => m [FilePath]

Re-exporting convenient stuff

class MonadIOC m => MonadIO m where

Methods

liftIO :: IO a -> m a

gmlGetSession :: (MonadIO m, MonadBaseControl IO m) => GmlT m HscEnv

gmlSetSession :: (MonadIO m, MonadBaseControl IO m) => HscEnv -> GmlT m ()