Language.Haskell.GhcMod.Types

Documentation

type IOish m = (Functor m, MonadIO m, MonadBaseControl IO m, ExceptionMonad m)

A constraint alias (-XConstraintKinds) to make functions dealing with GhcModT somewhat cleaner.

Basicially an IOish m => m is a Monad supporting arbitrary IO and exception handling. Usually this will simply be IO but we parametrise it in the exported API so users have the option to use a custom inner monad.

type MonadIOC m = MonadIO m

class MonadIOC m => MonadIO m where

Methods

liftIO :: IO a -> m a

data OutputStyle

Output style.

Constructors

LispStyle
S expression style.

PlainStyle
Plain textstyle.

newtype LineSeparator

The type for line separator. Historically, a Null string is used.

Constructors

LineSeparator String

data FileMapping

Constructors

FileMapping

Fields

fmPath :: FilePath

fmTemp :: Bool

type FileMappingMap = Map FilePath FileMapping

data ProgramSource

Constructors

ProgramSourceUser

ProgramSourceStack

data Programs

Constructors

Programs

Fields

ghcProgram :: FilePath

ghc program name.

ghcPkgProgram :: FilePath

ghc-pkg program name.

cabalProgram :: FilePath

cabal program name.

stackProgram :: FilePath

stack program name.

data OutputOpts

Constructors

OutputOpts

Fields

ooptLogLevel :: GmLogLevel

Verbosity

ooptStyle :: OutputStyle

ooptLineSeparator :: LineSeparator

Line separator string.

ooptLinePrefix :: Maybe (String, String)

Stdout/err line multiplexing using prefix encoding. fst is stdout, snd is stderr prefix.

data Options

Constructors

Options

Fields

optOutput :: OutputOpts

optPrograms :: Programs

optGhcUserOptions :: [GHCOption]

GHC command line options set on the ghc-mod command line

optOperators :: Bool

If True, browse also returns operators.

optDetailed :: Bool

If True, browse also returns types.

optQualified :: Bool

If True, browse will return fully qualified name

optHlintOpts :: [String]

optFileMappings :: [(FilePath, Maybe FilePath)]

defaultOptions :: Options

A default Options.

data Project

Constructors

CabalProject

SandboxProject

PlainProject

StackProject StackEnv

isCabalHelperProject :: Project -> Bool

data StackEnv

Constructors

StackEnv

Fields

seDistDir :: FilePath

seBinPath :: [FilePath]

seSnapshotPkgDb :: FilePath

seLocalPkgDb :: FilePath

data Cradle

The environment where this library is used.

Constructors

Cradle

Fields

cradleProject :: Project

cradleCurrentDir :: FilePath

The directory where this library is executed.

cradleRootDir :: FilePath

The project root directory.

cradleTempDir :: FilePath

Per-Project temporary directory

cradleCabalFile :: Maybe FilePath

The file name of the found cabal file.

cradleDistDir :: FilePath

The build info directory.

data GmStream

Constructors

GmOutStream

GmErrStream

data GhcModEnv

Constructors

GhcModEnv

Fields

gmOptions :: Options

gmCradle :: Cradle

data GhcModOut

Constructors

GhcModOut

Fields

gmoOptions :: OutputOpts

gmoChan :: Chan (Either (MVar ()) (GmStream, String))

data GhcModLog

Constructors

GhcModLog

Fields

gmLogLevel :: Maybe GmLogLevel

gmLogVomitDump :: Last Bool

gmLogMessages :: [(GmLogLevel, String, Doc)]

data GmGhcSession

Constructors

GmGhcSession

Fields

gmgsOptions :: ![GHCOption]

gmgsSession :: !(IORef HscEnv)

data GhcModCaches

Constructors

GhcModCaches

Fields

gmcPackageDbStack :: CacheContents ChCacheData [GhcPkgDb]

gmcMergedPkgOptions :: CacheContents ChCacheData [GHCOption]

gmcComponents :: CacheContents ChCacheData [GmComponent GMCRaw ChEntrypoint]

gmcResolvedComponents :: CacheContents [GmComponent GMCRaw (Set ModulePath)] (Map ChComponentName (GmComponent GMCResolved (Set ModulePath)))

data GhcModState

Constructors

GhcModState

Fields

gmGhcSession :: !(Maybe GmGhcSession)

gmComponents :: !(Map ChComponentName (GmComponent GMCResolved (Set ModulePath)))

gmCompilerMode :: !CompilerMode

gmCaches :: !GhcModCaches

gmMMappedFiles :: !FileMappingMap

data CompilerMode

Constructors

Simple

Intelligent

defaultGhcModState :: GhcModState

data GhcPkgDb

GHC package database flags.

Constructors

GlobalDb

UserDb

PackageDb String

type GHCOption = String

A single GHC command line option.

type IncludeDir = FilePath

An include directory for modules.

type PackageBaseName = String

A package name.

type PackageVersion = String

A package version.

type PackageId = String

A package id.

type Package = (PackageBaseName, PackageVersion, PackageId)

A package's name, verson and id.

pkgName :: Package -> PackageBaseName

pkgVer :: Package -> PackageVersion

pkgId :: Package -> PackageId

showPkg :: Package -> String

showPkgId :: Package -> String

newtype Expression

Haskell expression.

Constructors

Expression

Fields

getExpression :: String

newtype ModuleString

Module name.

Constructors

ModuleString

Fields

getModuleString :: String

data GmLogLevel

Constructors

GmSilent

GmPanic

GmException

GmError

GmWarning

GmInfo

GmDebug

GmVomit

type PkgDb = Map Package PackageConfig

Collection of packages

data GmModuleGraph

Constructors

GmModuleGraph

Fields

gmgGraph :: Map ModulePath (Set ModulePath)

data GmComponentType

Constructors

GMCRaw

GMCResolved

data GmComponent t eps

Constructors

GmComponent

Fields

gmcHomeModuleGraph :: GmModuleGraph

gmcName :: ChComponentName

gmcGhcOpts :: [GHCOption]

gmcGhcPkgOpts :: [GHCOption]

gmcGhcSrcOpts :: [GHCOption]

gmcGhcLangOpts :: [GHCOption]

gmcRawEntrypoints :: ChEntrypoint

gmcEntrypoints :: eps

gmcSourceDirs :: [FilePath]

data ModulePath

Constructors

ModulePath

Fields

mpModule :: ModuleName

mpPath :: FilePath

data GhcModError

Constructors

GMENoMsg
Unknown error

GMEString String
Some Error with a message. These are produced mostly by fail calls on GhcModT.

GMECabalConfigure GhcModError
Configuring a cabal project failed.

GMEStackConfigure GhcModError
Configuring a stack project failed.

GMEStackBootstrap GhcModError
Bootstrapping stack environment failed (process exited with failure)

GMECabalCompAssignment [(Either FilePath ModuleName, Set ChComponentName)]
Could not find a consistent component assignment for modules

GMEProcess String String [String] (Either Int GhcModError)
Launching an operating system process failed. Fields in order: function, command, arguments, (stdout, stderr, exitcode)

GMENoCabalFile
No cabal file found.

GMETooManyCabalFiles [FilePath]
Too many cabal files found.

GMEWrongWorkingDirectory FilePath FilePath

lGmcResolvedComponents :: forall cat. ArrowApply cat => Lens cat GhcModCaches (CacheContents [GmComponent GMCRaw (Set ModulePath)] (Map ChComponentName (GmComponent GMCResolved (Set ModulePath))))

lGmcPackageDbStack :: forall cat. ArrowApply cat => Lens cat GhcModCaches (CacheContents ChCacheData [GhcPkgDb])

lGmcMergedPkgOptions :: forall cat. ArrowApply cat => Lens cat GhcModCaches (CacheContents ChCacheData [GHCOption])

lGmcComponents :: forall cat. ArrowApply cat => Lens cat GhcModCaches (CacheContents ChCacheData [GmComponent GMCRaw ChEntrypoint])

lGmMMappedFiles :: forall cat. ArrowApply cat => Lens cat GhcModState FileMappingMap

lGmGhcSession :: forall cat. ArrowApply cat => Lens cat GhcModState (Maybe GmGhcSession)

lGmComponents :: forall cat. ArrowApply cat => Lens cat GhcModState (Map ChComponentName (GmComponent GMCResolved (Set ModulePath)))

lGmCompilerMode :: forall cat. ArrowApply cat => Lens cat GhcModState CompilerMode

lGmCaches :: forall cat. ArrowApply cat => Lens cat GhcModState GhcModCaches

lOptQualified :: forall cat. ArrowApply cat => Lens cat Options Bool

lOptPrograms :: forall cat. ArrowApply cat => Lens cat Options Programs

lOptOutput :: forall cat. ArrowApply cat => Lens cat Options OutputOpts

lOptOperators :: forall cat. ArrowApply cat => Lens cat Options Bool

lOptHlintOpts :: forall cat. ArrowApply cat => Lens cat Options [String]

lOptGhcUserOptions :: forall cat. ArrowApply cat => Lens cat Options [GHCOption]

lOptFileMappings :: forall cat. ArrowApply cat => Lens cat Options [(FilePath, Maybe FilePath)]

lOptDetailed :: forall cat. ArrowApply cat => Lens cat Options Bool

lOoptStyle :: forall cat. ArrowApply cat => Lens cat OutputOpts OutputStyle

lOoptLogLevel :: forall cat. ArrowApply cat => Lens cat OutputOpts GmLogLevel

lOoptLineSeparator :: forall cat. ArrowApply cat => Lens cat OutputOpts LineSeparator

lOoptLinePrefix :: forall cat. ArrowApply cat => Lens cat OutputOpts (Maybe (String, String))

lStackProgram :: forall cat. ArrowApply cat => Lens cat Programs FilePath

lGhcProgram :: forall cat. ArrowApply cat => Lens cat Programs FilePath

lGhcPkgProgram :: forall cat. ArrowApply cat => Lens cat Programs FilePath

lCabalProgram :: forall cat. ArrowApply cat => Lens cat Programs FilePath

data ModuleName :: *

A ModuleName is essentially a simple string, e.g. Data.List.

mkModuleName :: String -> ModuleName

moduleNameString :: ModuleName -> String